removeProperties: update to new environment
The old startSES shim was dealing with a very different environment (ES5),
and was trying to report problems that it encountered instead of merely
failing. We can simplify it now.

The outer function was rewritten to modify an argument named 'global'.

Also update to modern things like 'const' and 'let' instead of using 'var'.