rewrite fixups as injected strings, change s.confine to SES.confine
s.confine was a method on a frozen SES realm, allowing someone *outside* to
evaluate code *inside*.

In contrast, the new SES.confine is only available from *inside* a frozen SES
realm, and evaluates in the same context as the caller. It maps to the
parent's r.evaluate . After ten years of work on EcmaScript, this finally
provides a safe two-argument eval (taking code and endowments, using
primordials but not anything else from its global environment).