build-intermediate: remove async, let it work on node >=v6
Using 'async function' in a build-time tool was preventing test (or anything)
from working on older versions of Node which don't provide support by
default.

This allows tests to *run* on v6 and newer, although they're only passing on
v7 and newer (TypeError: getOwnPropertyDescriptors is not a function in the
Realm's evaluators.js line 32).