add SES.confineExpr, useful for turning function strings into functions
This just adds a pair of parenthesis around the string being evaluated, so
you can start with a function declaration in one Realm, use the built-in
stringification tools to make it into a string, then evaluate that string to
get back the same function body in a different Realm:

```
 function foo(a) {
   return a+1;
 }
 const body = `${foo}`;

 const f = SES.confineExpr(body);
 f(2); // === 3

 // equivalent to SES.confine(`(${foo})`)
```

Of course this is most useful when one domain stringifies the function, and a
different one evaluates it.

SES.confine evaluates the string as a program in a temporary global scope. It
returns the "completion value", which is the last expression of the program.
Function declarations aren't expressions, so if you didn't add the
parenthesis, the evaluation would *define* `foo()` in the temporary scope,
but the function would be lost when the global scope disappears.

Adding the parenthesis turns the function declaration into a function
definition expression, which *does* serve as a completion value.

To make this work without the parens, you'd have to make the last statement
of the evaluated program be the name of the function, which isn't very
natural and doesn't enable easy stringification:

```
 function foo(a) {
   return a+1;
 }
 foo
```