src/index: fix exports
Commit 15ecf152 added non-default exports ('def' and 'Nat'), which changed
the way rollup creates the bundle (dist/ses-shim.js), which broke the demo:
it wanted SES.makeSESRootRealm(), but would have needed to use
SES.default.makeSESRootRealm() or SES.SES.makeSESRootRealm().

The demo page kept working because we forgot to re-build the bundle after
making that change, so it kept using the older bundle.

This patch synthesizes an object for export that allows both
import-of-default and import-of-SES to work. It also adds the
--output.exports=named flag to hush a warning about doing this sort of mixed
export. If we had listened to that warning earlier, we might have noticed the
breakage.

I'll rebuild the bundle in the next commit.