tame Math.random with an exception, rather than NaN
We used NaN in Date.now() because we found that the moment.js library was
calling Date.now() but not using the result, so it crashed even when nobody
was using the tamed API surface. We're guessing that this won't happen with
Math.random, and in general it's better to get an error than a weird
numerical result (even though we'd be within our rights to just return e.g.
"4" every time.. how can you tell it's not actually random??)