use whitelist.js and anonIntrinsics to remove non-approved properties
createSES: stabilize name of getAnonIntrinsics

Our removeProperties() function was depending upon getAnonIntrinsics() being
available by that name. We took the source of getAnonIntrinsics() and
evaluated it just before evaluating removeProperties(), but rollup (and
probably other bundlers) might mangle the names of functions when it sees
multiple instances of them. In this case, the function was actually named
getAnonIntrinsics$1, which made removeProperties() break.

This change forces getAnonIntrinsics() to be known by that exact name, even
if rollup changes the name in the original function declaration.