switch back to safer traversal algorithm
The deepFreeze.js from SES (originally from Caja) was careful to freeze an
object before following its traverseable properties, to guard against sneaky
objects that use a Proxy or accessor function to change their properties
during freezing.

Our previous version made the promise that when TypeError is thrown (due to
one of the reachable objects having a prototype that was not already in the
root set), no freezing would take place: freezing would be all-or-nothing,
either success (everything frozen) or error (nothing frozen).

These two goals are mutually incompatible.

We decided to have harden() freeze everything that is reachable, and then do
the prototype-is-in-rootSet check at the end. So you still get a TypeError if
you haven't previously handled the prototypes of everything reachable, but
you no longer get to rely upon the objects *not* being frozen.

You don't get to rely upon them being frozen either: while that's the
behavior of the code right now, we might not want to commit to that, so for
now the API contract says that if you get TypeError, the reachable objects
may or may not be frozen. TypeError is not intended to be a recoverable
error.