remove Nat from global, use require('@agoric/nat') to get it
In the longer run, Nat will be removed from SES entirely, and a more generic
approach to require() will allow e.g. PlaygroundVat to provide it as a global
to imported code. But for now, changing this name to match the NPM package
name should let us start writing code that works in both modes.