anonIntrinsics: add AsyncIteratorPrototype
We were missing AsyncIteratorPrototype, which is the
prototype (getPrototypeOf) of AsyncGeneratorPrototype, which is
the (dot)prototype of AsyncGenerator. Our current deepFreeze follows both
sorts of links, so AsyncIteratorPrototype was being frozen, but the new
version (harden) doesn't follow getPrototypeOf, so it wouldn't have been
frozen there.

Also add AsyncIteratorPrototype to the whitelist so it will be retained.