buildTable: add anonIntrinsics to the fringe
We weren't including these earlier, so harden() would fail when given an
async function, a generator, or an async generator.

Also clean up the table-building process, to avoid including the global
object in the fringe.

Rename functions for clarity, and change buildTable.js into an ES6 module to
match everything else in the package (and avoid an import problem in which
the whitelist was converted into something with a `.default` property).