integration test: tell parcel to stop polyfilling so aggressively
The default settings for parcel are to polyfill enough features to work in
99.75% of browsers. Unfortunately this means generators (`function*`) are
rewritten, and this rewriting depends upon a plugin that we aren't including:
we get errors like `ReferenceError: regeneratorRuntime is not defined`.

This lowers our target browser support level to 50%, which apparently stops
rewriting generators.

The cleaner fix for this would be to specifically disable generator
rewriting, but in practice anyone using parcel is going to need to make sure
their rewriting instructions match the plugins they provide, so there's not
much more we can do in our testing framework to help out.