update rollup config to find main.js
Many (all?) of these packages were converted to `type: "module"`, and their
`index.js` removed in favor of a `main.js`. This updates the
`rollup.config.js` to look for this new filename.. without it, a `yarn build`
fails when rollup cannot find the old `index.js` files.

It also shuffles the top-level `package.json` to build `make-hardener` before
`harden`, and adds explicit devDependencies on rollup (and the node-resolve
plugin) to both packages.

refs #6