SES: tolerate RegExp() without new, reenable tests
Plain `RegExp()` was failing with an `undefined is not a constructor` error,
which broke a lot of packages unless `noTameRegExp: true` was set. The
test262 tests which happened to exercise this case were disabled (probably
for other errors).

Thanks to @erights for the recommended patch. I reenabled 21 test262 tests
that now pass with this fix.

closes #230