chore: update yarn.lock with newer 'acorn'
The transitive closure of dependencies causes the SES-shim tree to depend
upon three distinct versions of the `acorn` package (5.x, 6.x, 7.x). Our
yarn.lock was previously selecting 7.1.0 to satisfy the last one, but that
version has a security advisory, which is addressed by 7.1.1.

This changes (only) yarn.lock to use 6.4.1 instead of 6.4.0, and 7.1.1
instead of 7.1.0. These versions seem to be compatible with all the
constraints expressed by our dependency set, and are (currently) free of
security alerts.