Add deferred module exports namespace proxy
A ModuleNamespace is a wrapper for the ModuleNamespace that it will become after it has been linked.  Having a ModuleNamespace type allows a Compartment to give users a handle to the exports of a module before it has been loaded, analyzed, or linked.  This handle can be used to represent linkage to that module when it is passed to a Compartment through its module map argument.  The association is preserved behind the scenes by WeakMaps both from compartments to namespaces and namespaces back to compartments.

TODO pass through key iteration order and figure out what other
properties should be preserved by the module exotic object.