Add private module instance type (#283)
This function implements the instantiation convention for a ModuleStaticRecord.  The module static record is a function and metadata about what a module imports and exports, and how it presents those exports based on whether they are mutable and whether they is a temporal-dead-zone between when the module is instantiated and when an exported field becomes accessible.  The module instance type connects the module to namespace of imported and exported modules and provides the mechanism for eventually executing that module.

This is a slightly altered copy of the module instance constructor from the `@agoric/make-importer` package. The changes facilitate coupling with compartments. We may elect to project these changes back out and take a dependency.

The facilities for threading additional endowments have been removed since Compartments only allow thread endowments from their constructor and as arguments to one-off evaluate calls.  The instance constructor also accounts for the presence of a compartment on the linkage record instead of receiving an evaluator through its own constructor.  The signature of compartment.evaluate also differs from the previously assumed evaluator and that has been accounted for.