chore(bundle-source): perform tildot transform during bundling
Our vats and contracts (can be) written with "tildot syntax": bob~.foo(arg1,
arg2). This must be transformed before it can be executed, into
HandledPromise.applyMethod(bob, "foo", [arg1, arg2]).

Currently, the tildot transform is applied by the kernel by adding a
transform function to the evaluate() options. The controller gets this
transform from the default-evaluate-options, and currys it into the evaluator
that it provides to the kernel (the metering transform is also bundled in
here).

We're moving the responsibility for the tildot transform from the kernel into
the bundle-source process. Tildot transformation is "optional", in the sense
that we're ok with vats/contracts trying to dodge it (unlike the injection of
metering, which must be enforced everywhere to be effective). So it's ok for
us to have users tildot-transform their code on their own machines, before
sending the bundled artifact up to the chain. If they skip this step, their
code will fail to parse when loaded into a vat/Compartment, but that's their
problem, not ours.

This adds the tildot transform into `bundle-source`. For a while this will be
redundant: we won't remove it from the kernel/controller until we're on
new-SES. Applying the transform twice shouldn't be a problem, the second time
will be a no-op.

closes #1101

thanks @michaelfig for the fixes