fix(ses)!: Simplify transforms
This change simplifies the compartment transforms for evaluated programs to just an array of functions that accept and return program strings.  The previous version of this API enabled transforms to add endowments and left room for future transformable properties.  We have elected to use the compartment `evaluate` function to introduce lexical endowments and `globalThis` to introduce enumerable and computable endowments.