feat(endo): Add search
A Node.js application is a module and its transitive dependencies.
To provide as close an approximation of Node.js, running an Endo
application is also the importing of an entry module.
To construct the compartment graph for the module, the search
function starts with a path to the module and searches for
the first parent directory that contains a "package.json".

Node.js does not require there to be a "package.json".
Neither do web applications using ESM or Deno applications.
This is a constraint that we might relax in a future version by creating
a special compartment for entry modules.