ses(fix): Repair defineProperty internally (#361)
* fix(ses): Downsize defineProperties to defineProperty in enablements
* fix(ses): Apply wisdom of commons homogeneously

Throughout SES, we should be using a patched version of defineProperty because the original may in some environments return false instead of throwing an error when applied to a window proxy object.

The defineProperties function behaves correctly in all situations.

Some of the modules within ses-shim use a commons.js module as the source of Object function properties.  Using this module homogeneously ensures that the patched version of defineProperty happens to be correct in all cases.