fix: callSite objects have properties on their prototype
v8 puts the various `getFunctionName` accessors on a callSite's *prototype*,
not on the callSite object itself. The source-map-support library, which
wraps callSite objects to augment their position-finding details, relies upon
being able to find these properties on that prototype.

Previously, when in `errorTaming: 'unsafe'` mode, we placed the
non-integrity-breaking accessor wrappers on the replacement callSite object.
This changes the taming code to put them on a new object, and uses that
object as the prototype of the one we give to Error.prepareStackTrace.