fix: reveal more callSite methods to Error.prepareStackTrace
The `source-map-support` library, used by the `tap` test framework (and
others), depends upon additional properties of the callSite object. I'm
pretty sure these are no worse than the other properties we're currently
exposing (they violate confidentiality of everyone in the stack, but not
integrity).

This wraps and exposes the following properties to `Error.prepareStackTrace`
when `errorTaming` is `unsafe`:

* isToplevel
* isNative
* getPosition
* getScriptNameOrSourceURL