expose more callSite properties, put them on the prototype (#369)
* fix: reveal more callSite methods to Error.prepareStackTrace

The `source-map-support` library, used by the `tap` test framework (and
others), depends upon additional properties of the callSite object. I'm
pretty sure these are no worse than the other properties we're currently
exposing (they violate confidentiality of everyone in the stack, but not
integrity).

This wraps and exposes the following properties to `Error.prepareStackTrace`
when `errorTaming` is `unsafe`:

* isToplevel
* isNative
* getPosition
* getScriptNameOrSourceURL

* fix: callSite objects have properties on their prototype

v8 puts the various `getFunctionName` accessors on a callSite's *prototype*,
not on the callSite object itself. The source-map-support library, which
wraps callSite objects to augment their position-finding details, relies upon
being able to find these properties on that prototype.

Previously, when in `errorTaming: 'unsafe'` mode, we placed the
non-integrity-breaking accessor wrappers on the replacement callSite object.
This changes the taming code to put them on a new object, and uses that
object as the prototype of the one we give to Error.prepareStackTrace.

* chore: add new tap-vs-lockdown tests

Most of our tests use the "tap" library to exercise various parts of the SES
codebase. They `import 'tap'`, then call `lockdown()` with some options, then
examine the primordials which `lockdown()` has modified.

But "tap" itself modifies the primordials, by using libraries like
"source-map-support", which augments `Error` and `Error.prepareStackTrace` to
display stack traces that follow sourcemaps. When our test files `import
'tap'` first, then call `lockdown()`, SES is wrapping the results of
`source-map-support`.

Downstream developers will do this the other way around: they will import a
module with a name like `install-ses`, which calls the SES `lockdown()` at
the top level, after which they will `import 'tap'`, and after *that* they'll
import their own code, and then run tests. In that environment,
`source-map-support` is operating on the wrappers produced by `lockdown()`.
And that environment didn't always work (see #367 for details).

This adds some basic tests which first call `lockdown()`, *then* imports
`tap`. We can't exercise the most interesting cases (which are when a tap
assertion fails, and tap attempts to display a stack trace to the failure),
because those would cause the SES tests to fail. But we can at least make
sure the success cases work. And if a problem is suspected, a SES developer
can make a small temporary edit to these new test cases and see what happens
when an assertion fails.

refs #367