ses(chore): Address moduleMap on-demand
Prior to this change, the Compartment constructor translated the module map into aliases.  After this change, the compartment loader addresses the module map on demand.  Some validation remains in the constructor for early errors when possible.  This change prepares the Compartment for a dynamic module map hook, necessary to emulate wildcard linkage for Node.js packages, and also sufficient to enable cyclic dependencies among modules in separate compartments.