feat(ses): Revert "Export SES Transforms (#608)"
SES is constrained to use a regular expression for evasive transforms.
SES exposes the evasive transforms as options to `evaluate`.
For cases where a bundler is involved, this problem is better solved
with a parser transform that can distinguish comments from names in
scope and can discern whether the transformation it proposes is likely
to cause semantic breakage.

This reverts commit 5ec8858648254e7cd2a1bf9c054a1d5d2749c31b.