feat(endo): Freeze all global objects (#631)
By default, freeze the global object in every compartment after endowments. We may in a future version allow some compartments to remain thawn through explicit configuration from the application (possibly recommended on a per library basis). We may in a future version provide a hook for alternate global post processing like `harden`.  We do not take a dependency on `harden` because that would limit the use of compartment mapper as a demonstration of how Node.js could use compartments as a standard module loader API, orthogonal to SES.