feat(compartment-mapper): Developer dependencies
The import and archive functions now accepta `dev` flag that indicates that the `devDependencies` of the entry module's package should be included in the compartment graph.  This must be passed explicitly and is off by default since that will be the norm for applications.

This is the bug of the day blocking https://github.com/Agoric/agoric-sdk/pull/3273

In furtherance of https://github.com/Agoric/agoric-sdk/issues/2684