refactor(ses): Consolidate declarations of uncurryThis and related functionality into commons.js (#821)
Some functions that appear as though they should be exported from `commons.js` were declared in `whitelist-intrinsics.js`. `objectHasOwnProperty` and `ownKeys` (from `Reflect`) are now imported from `commons.js` instead of declared in `whitelist-intrinsics.js`. As a consequence, `uncurryThis` is no longer used in `whitelist-intrinsics.js`, and that function is completely removed from that file. `ownKeys` is added as an export of `commons.js` along with the other properties of `Reflect` exported from that file.