fix(promise-kit): make strict typing compliant (#3397)
## Description

`@agoric/promise-kit` doesn't specify a `types` entry in its `package.json` causing `tsc` to parse its `.js` source to extract jsdoc typing. This parsing would also validate this internal implementation, which would raise errors when the consumer of the package is in `strict` mode (mostly `noImplicitAny`).

### Security Considerations

There is a single runtime change from `p.domain` to `'domain' in p` which is actually more correct.

### Documentation Considerations

N/A

### Testing Considerations

We should enable `"strict": true` for this package's `jsconfig.json` but that currently isn't possible because `global.HandledPromise` is only declared in `@agoric/eventual-send` which is not imported directly by `promise-kit`.