fix(static-module-record): Do not duplicate comments (#863)
This fixes a problem with static module records duplicating comments.  The problem arises from running a babel transform on the AST produced by a prior transform, which in our case causes the trailingComments of one node to be equal but not identical to the leadingComments of a subsequent node, and the Babel codegen interpreting these as separate.  The effect is that the retainLines flag causes the codegen to attempt to fall back to its intended line number by folding subsequent lines.

This PR uses the TDD convention of posting a failing test followed by the code that fixes it, so the test can be tested by checking out the first commit and running the failing test, then verifying it passes on the second.

- test(static-module-record): Test comment duplication
- fix(static-module-record): Comment duplication