fix(ses): Direct eval check should not preclude no-eval under CSP (#1004)
This change adds nuance to the code that checks for whether eval is
direct-eval. We do want to screen environments where eval is indirect.
But, if eval throws an exception instead, that would indicate a
content-security-policy that excludes unsafe-eval, and SES should be
usable for this case in a future version.

Currently, this case fails for other reasons.
So this change adds a known-failing test.