fix(compartment-mapper): Needless genericity considered harmful
The has utility function in the compartment mapper did not need to be generic.  Because it was generic, at a vast distance down the dependency chain, TypeScript proposed bizarre combinations of types for each of the uses, probably sampling the first ambient type that emitted a valid function, then failed type check in compartment-mapper.