feat(stream)!: Protect streams against reentrance
This is a proposed change that would by default protect stream methods from reentrance by defering method invocations to subsequent turns.  However, it does not work because eventual-send hardens arguments, and for streams, sending a typed array must be a valid interaction.

I propose that harden needs to treat typed arrays as a special case of collections like Map and Set, which can be hardened. The typed arrays are equally mutable, in that their content is mutable. Hardening a typed array could instead prevent extensions and make its non-content properties non-configurable.