fix(eventual-send): hoist closures to discourage argument retention
We've seen HandledPromise invocation arguments be retained against GC
for surprisingly long periods of time. It seems that the closures
defined in trackTurns() were holding on to more of their lexical scope
than strictly necessary (a lot of JS engines do this). This hoists
those closures up to top-level functions to discourage that retention.

closes #1245