A .then detection added.
This is usefull for detecting intrest in what a remote promise resolves to in CapTP, for example. In that case an op:listen or such can be sent over the connection and, crucially, skip that iff no interest is shown and the handled promise is only used as an target or an argument of an eventual send.