fix(typescript-eslint): workaround for `exports.js` with `.d.ts`
The problem is that typescript configs containing both `*.js` and
`*.ts` files are confused when a `foo.js` file occurs next to a
`foo.d.ts` file.  That makes typescript-eslint produce:

https://typescript-eslint.io/linting/troubleshooting/#i-get-errors-telling-me-eslint-was-configured-to-run--however-that-tsconfig-does-not--none-of-those-tsconfigs-include-this-file

A cheap workaround for now is just to remove `exports.js` from
Endo's typescript-eslint configuration, but that does not prevent
the error after, say `yarn prepack` when there are hordes of
`.d.ts` files next to the `.js` files.  Too bad.