fix(ses): fix ThirdPartyStaticModuleInterface type
The `proxiedExports` param to the `execute` function cannot be `Object`, which is the literal `Object`.  It must be _some_ kind of `Object`, however, a `Record` is suitable.

However, the property can be a `PropertyKey` if the exports come from CommonJS and can only be a `string` if the exports come from ESM. Use `string` for now and we can widen the type later.