feat(cli,daemon): Support dot-delimited petname paths in eval
Adds support for dot-delimited petname paths to the CLI
`eval` command. This required modifications to the `evaluate`
method of the daemon's `host.js`, and the introduction of a
new formula type, `lookup`, modeled on the `web-bundle`
formula. Also adds a unit test for this behavior.

The `lookup` formula type is necessary because the names
in a lookup path may have neither petnames nor formula
identifiers associated with them. Consider:

```text
> endo eval '10' --name ten
10
> endo eval 'foo' foo:INFO.ten.source
10
```

The only way retrieve the value for `source` is to call
`E(HOST).lookup('INFO', 'ten', 'source')`, and since the `eval`
formula expects its values to be mediated by formula identifiers,
the lookup of `INFO.ten.source` must itself be stored as
a formula.