refactor(daemon): Replace "agent" language with "hub" for lookup
Lookup formulas must be associated with the namespace for
which the pet name path is valid. In practice, this means
an object with a variadic `lookup()` method. Previously, these
objects were referred to as "agents", but in the future, not
all such object will be agents (i.e. guests or hosts). Therefore,
here we replace all references to "agent" with "hub", short
for "naming hub".