fix(daemon): Correctly use worker formula number in incarnateEval
`incarnateEval()` was accidentally passing worker formula identifiers
instead of formula numbers when it received a specified worker formula
identifier. This caused `eval` formulas to always fail on Windows due to
a `:` in the path.

Surprisingly, fixing this caused the "indirect termination" test to
fail. This is likely just surfacing a deeper problem, see #2074.