refactor(daemon): Synchronize incarnateReadableBlob()
Merges the daemon's `storeReaderRef()` into `incarnateReadableBlob()`
and synchronizes the implementation. Also converts `makeReadableBlob()`
into `makeControllerForReadableBlob()` to align it with the family of
functions that already exist for most other formulas.

Contrary to its predecessor, `incarnateReadableBlob()` returns the value
for stored value, namely a `FarEndoReadable`. A unit test is added for
  the case of immediately using a stored value without naming it.