feat(daemon): Add directory `locate()` method
Adds `locate()` to the directory, which accepts a pet name and returns a
locator string. `locate()` will succeed for any named local value at any
point in the daemon's lifecycle.

Adds the `remote` formula type to populate the `type` query param of
locators, since the "real" type of remote values cannot be known.