feat(daemon): Add bidirectional multimap
Adds makeBidrectionalMultimap() alongside other multimap factories. This
abstraction is intended to replace the use of multiple distinct maps in
the pet store module.

In its current implementation, the map cannot weakly hold on to its
values, because it needs strong references in order to enumerate them.
(Recall that WeakSets have no facilities for enumerating their values.)